Hide ValueSetsReachingValue: implementation detail
This class is used to implement the valueSetsReachingValue() function,
but callers to that function do not need to know anything at all about
this class.  They don't even need to know that it exists.
